requirements.txt
uvicorn==0.34.1
fastapi==0.115.12
requests==2.32.3
msal==1.32.0
apscheduler==3.11.0
pymongo==4.12.0
email-validator==2.2.0
python-dotenv==1.1.0